codegen: Make deterministic register allocation across different runs
The range loop traverses a map in a non-deterministic order, and the
register index returned by GetReg() relied on this behavior leading to
different ordering of register allocation in the generated assembly
across different runs (though the meaning was conserved). To avoid this
we convert the O(1) algorithm to O(RegLimit) algorithm, and since
RegLimit is known it is still O(1) but with slightly larger number of
(constant) iterations than the previous version.