Move the logic of text segment updation when spilling inside "GetReg()"
The goal of 227d43d was to reduce the redundances resulting from the
updates required to be made due to side-effects induced by a call to
"GetReg()" (updating lookup tables, text section etc). This commit is
one step ahead towards that goal.

Ideally, the text and data segment data structures (TextSec & DataSec)
should be defined inside "codegen.go" but for this sole purpose they had
to be moved inside "tac.go" to prevent an import cycle.