Fix register spilling during allocation
GetReg is updated to delete all the lookup entries for the spilled
variables (as opposed to earlier when the dst variable's register was
being skipped). In case the src variable is spilled and its register is
given to dst, the src variable's lookup entry is temporarily marked.
This entry is then deleted afterwards on the caller's end (codegen.go)
when the relevant statement is pushed into the text segment DS. This
ensures proper cleanup of all the lookup tables.

The problem arised when the lookup table was not being properly updated
corresponding to the spilled src variable, which lead to it not being
loaded from memory into a new register on its next use.

Signed-off-by: Sunil Kumar Pandey <sunilpandey.pandey09@gmail.com>
Signed-off-by: Shivansh Rai <shivansh@freebsd.org>