Avoid using global variable to check if return stmt is to be inserted
Instead use the length of the defer stack. If the stack is empty then
the return stmt is inserted as usual in the production rule for
ReturnStmt, otherwise it is deferred until the end of the function block
when all the deferred function calls are inserted in the IR and the
stack is empty.