Fix register allocation
Previously only temporaries ($t0 - $t4) were being allocated to
variables, the reason being that register spilling was much more
apparent when dealing with a small number of registers. This is updated
to make all the 32 registers available for allocation. However, $1 is
reserved by the assembler for pseudo instructions and hence is not
assigned to variables. This fixes the frequent register spilling which
was being observed previously.

Add a note explaining how register spilling works currently.

Add a TODO file with tasks to be done in the near future.

NOTE: This commit might break code generated for arrays.