tac: Update logic for basic block creation on encountering a jump
If the variables belonging to a branch instruction haven't been
allocated a register, an allocation is done. Since the branch
instruction marks the end of the current basic block, a store is
performed for all the variables which have been allocated a register. If
the block only contains a label statement, the branch instruction will
invoke a load for its variables, and immediately after this a store will
be performed for these variables. This store instruction is unnecessary
since the value of branch variables wasn't modified in this block.
Therefore if such a case arises, avoid creating a new basic block.

For a demo of the unnecessary load/store, consider the following IR -

	label, while
	ble, exit, n, 0
	/, n, n, 10

Generated assembly -

	while:
	        lw      $3, n
	        # Store variables back into memory
	        sw      $3, n
	        ble     $3, 0, exit     # exit -> $0

	        lw      $3, n
		div     $3, $3, 10      # n -> $3

The store and load instructions performed before and after ble
respectively are unnecessary.

tests: Update generated assembly