codegen: Update register allocation to be deterministic
When initializing the register priority queue, a higher priority is
given to registers with lower index so that they will be popped first.
This will result in a deterministic register allocation. In case all the
registers have their NextUse initialized to MaxInt (as was the case
previously), Pop() returns one non-deterministically.

This also fixes the travis build as the generated assembly does not vary
across different runs.